fix: 🐛 move datastore utils to th datasotre
✅ Closes: #187
