+++ src/core/carstore/repo/datastore/index.ts
@@ -18,7 +18,7 @@
  *  limitations under the respective licenses.
  ********************************************************************************/
 import { DataStore } from "@unipackage/datastore"
-import { ValueFields } from "@unipackage/utils"
+import { ValueFields, Result } from "@unipackage/utils"
 import { Car, CarReplica } from "../../types"
 import {
     CarDocument,
@@ -27,6 +27,14 @@ import {
     CarReplicaSchema,
 } from "./model"
 import { MongooseDataStore, DatabaseConnection } from "@unipackage/datastore"
+import { MatchingTarget } from "../../../../module/matching/target/types"
+import {
+    convertToCarArray,
+    convertToCarReplicasArray,
+} from "../../../../shared/converters"
+import { CarstoreEvm } from "../evm"
+import { DatasetRequirementEvm } from "../../../../module/dataset/requirement/repo/evm"
+import { DatasetProofs } from "../../../../module/dataset/proof/types"
 
 /**
  * Class representing a MongoDB datastore for Car entities.
@@ -51,7 +59,38 @@ export class CarMongoDatastore extends DataStore<
             )
         )
     }
+
+    /**
+     * Stores cars in the Carstore contract and updates the dataset requirement in the DatasetRequirement contract.
+     * @param options - Object containing necessary parameters.
+     * @param options.carstoreEvm - Instance of the CarstoreEvm contract.
+     * @param options.requirementEvm - Instance of the DatasetRequirementEvm contract.
+     * @param options.proofs - DatasetProofs containing proofs for car-related data.
+     * @returns A Promise resolving to a Result indicating the success or failure of the operation.
+     */
+    async storeCars(options: {
+        carstoreEvm: CarstoreEvm
+        requirementEvm: DatasetRequirementEvm
+        proofs: DatasetProofs
+    }): Promise<Result<any>> {
+        const cars = await convertToCarArray({
+            carstoreEvm: options.carstoreEvm,
+            requirementEvm: options.requirementEvm,
+            proofs: options.proofs,
+        })
+        for (let i = 0; i < cars.length; i++) {
+            const ret = await this.CreateOrupdateByUniqueIndexes(cars[i])
+            if (!ret.ok) {
+                return {
+                    ok: false,
+                    error: new Error(`storeCars error:${ret.error}`),
+                }
+            }
+        }
+        return { ok: true, data: cars }
+    }
 }
+
 /**
  * Class representing a MongoDB datastore for CarReplica entities.
  * Extends the DataStore class with CarReplica and CarReplicaDocument.
@@ -74,5 +113,27 @@ export class CarReplicaMongoDatastore extends DataStore<
                 connection
             )
         )
+    }
+
+    /**
+     * Stores car replicas in the CarReplicaDatastore and updates the matching target.
+     * @param options - Object containing necessary parameters.
+     * @param options.target - MatchingTarget to be updated with the stored car replicas.
+     * @returns A Promise resolving to a Result indicating the success or failure of the operation.
+     */
+    async storeCarReplicas(options: {
+        target: MatchingTarget
+    }): Promise<Result<any>> {
+        const carReplicas = convertToCarReplicasArray(options.target)
+        for (let i = 0; i < carReplicas.length; i++) {
+            const ret = await this.CreateOrupdateByUniqueIndexes(carReplicas[i])
+            if (!ret.ok) {
+                return {
+                    ok: false,
+                    error: new Error(`storeCarReplicass error:${ret.error}`),
+                }
+            }
+        }
+        return { ok: true, data: carReplicas }
     }
 }
